library_qr_generation/requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.4.0
